Birds table + topnav update
Added Birds table to database that include some details about birds. These details are shown on the sighting details page. Also the birdname selection bar is now using this database instead of a plain csv file.

Updated javascript to change the color of active page in the top navigation bar.